Fix Cfg.Analysis.focus* behavior with JumpTo pred
Previously, we would only simplify a `JumpTo` predecessor of the focal
node if the `JumpTo` was the _only_ statement in the predecessor. Also,
we would remove all targets. Now, we only require the `JumpTo` be the
_last_ statement in the predecessor, and if the focal node has an
address, we try to keep it in the `targets` list of the `JumpTo`.

Also change the type of JumpTo to reflect the corresponding change in
binaryninja-haskell